Move Tests to c3po Module
Tests belong in the package contents itself.

Fixes #83
